requirements.txt
algoliasearch==1.9.1
algoliasearch-django==1.2.4
BeautifulSoup
beautifulsoup4==4.5.1
cffi==1.8.3
stripe
cryptography==1.5.3
cssutils
dj-database-url==0.4.1
dj-static
Django==1.10.3
django-allauth==0.28.0
django-bootstrap3==7.1.0
django-debug-toolbar==1.6
django-rest-swagger==0.3.10
django-sslify
django-toolbelt
djangorestframework==3.5.3
djrill==2.1.0
enum34==1.1.6
gunicorn==19.6.0
idna==2.1
ipaddress==1.0.17
jsonfield
Markdown==2.6.7
ndg-httpsclient==0.4.2
newrelic==2.72.1.53
oauthlib==2.0.0
psycopg2==2.6.2
pusher==1.6
pyasn1
pycparser==2.17
pynliner==0.7.2
pyOpenSSL==16.2.0
python-dateutil==2.6.0
python-openid
pytz==2016.7
PyYAML==3.12
requests
requests-oauthlib==0.7.0
six
sqlparse==0.2.2
static3==0.7.0
templated-emails
tweepy
unicodecsv
urllib3==1.19
wheel==0.30.0a0
whitenoise==3.2.2
django-webpack-loader==0.4.1
